requirements.txt
PyGithub~=1.58.2
django-bakery
django
python-decouple
Jinja2~=3.1.2
click~=8.1.3
livereload
python-dotenv~=1.0.0
pytest
PyYAML~=6.0
Markdown~=3.3.7